Modified app to use Outlet.
Modified routes.
Make App.jsx the parent of all other pages.
Navbar is now rendered in App.jsx.
Moved useFetchProducts() from Shop.jsx to App.jsx.
The Cart button now shows the number of items in the cart.
Clicking on an Add to Cart button adds a value to the Cart button.